User can input the name of the preference, min & max temp and location. Displayable via TextView when I press on display.
However, I need to find a way how to make the ListView receive the title of the preference and make it display the preference's detailed info when I click on it in the ListView and press Display. To do so, I need to find a way to blend in the listActivity class with the userActivity class, in regards of the addPrefs() and etc.