Finished three Activities: PreferencesActivity, listSavedPrefsActivity and PreferenceInfoActivity.
PreferencesActivity: Allows the user to add new preferences.
- Toast appears when at least a box is not filled or the max temperature is smaller than min temperature

listSavedPrefsActivity: Allows the user to look at the saved preferences. They can also delete it from the list by holding their finger on a preference.

PreferenceInfoActivity: Allows the user to view details of a specific preference after selecting one from the listView.

Two extra classes, Preference and GlobalModel, are used to complement the 3 activities (putting the name of the preference in an ArrayList, calling their info like weather type and temperature, ...)

Signed-off-by: minhdo <minhdo@metropolia.fi>